Function oak_grpc_utils::generate_grpc_code

pub fn generate_grpc_code(
    protos: &[impl AsRef<Path>],
    include: &[impl AsRef<Path>],
    options: CodegenOptions,
) -> Result<()>

Generate gRPC code from Protobuf using tonic library.

The path to the root repository must be passed as include. All paths to .proto files must be specified relative to this path. Likewise, all imported paths in .proto files must be specified relative to this path.